Jamba-Team
No longer should add -realmname when adding team names to the team list. This is going to help with ISBoxerboxer addon that adds character-realmname
will not help if already have character-reamname and you may need to reset the pofile still, But they should no longer keep getting added by ISBoxer.

Added a button to add party members to the team under on the GUI team display so you don't have to use the the command line. ( I did not even know this there.)